+++ IRMIS/src/table.js
@@ -4,6 +4,7 @@ import $ from "jquery";
 import { applyFilter } from './filter';
 import { getFieldName } from "./road";
 import { exportCsv } from "./exportCsv";
+import { datatableTranslations } from "./datatableTranslations";
 
 let table;
 let pendingRows = [];
@@ -22,7 +23,7 @@ document.addEventListener('estrada.roadManager.roadMetaDataAdded', (data) => {
 
 document.addEventListener('estrada.table.roadMetaDataUpdated', (data) => {
     const road = data.detail.road;
-    table.row(`#${road.id}`).data(road).draw();
+    table.row('#' + road.id).data(road).draw();
 });
 
 // when a filter is applied, update the filter id whitelist
@@ -218,25 +219,25 @@ export const columns = [
         visible: false,
     },
     {
-        title: window.gettext("Start Point (DMS)"),
+        title: gettext("Start Point (DMS)"),
         data: "startDMS",
         defaultContent: "",
         visible: false
     },
     {
-        title: window.gettext("End Point (DMS)"),
+        title: gettext("End Point (DMS)"),
         data: "endDMS",
         defaultContent: "",
         visible: false
     },
     {
-        title: window.gettext("Start Point (UTM)"),
+        title: gettext("Start Point (UTM)"),
         data: "startUTM",
         defaultContent: "",
         visible: false
     },
     {
-        title: window.gettext("End Point (UTM)"),
+        title: gettext("End Point (UTM)"),
         data: "endUTM",
         defaultContent: "",
         visible: false
@@ -248,7 +249,7 @@ function initializeDataTable() {
         columns.unshift({
             title: "",
             data: null,
-            render: r => `<a class="image pencil" href="#edit/${r.getId()}/location_type"></a>`,
+            render: r => '<a class="image pencil" href="#edit/' + r.getId() + '/location_type"></a>',
             orderable: false,
             className: "edit-col"
         });
@@ -259,7 +260,8 @@ function initializeDataTable() {
         rowId: ".getId()",
         // default order is ascending by: road code, link code, & link start chainage
         order: window.canEdit ? [[1, 'asc'], [3, 'asc'], [7, 'asc']] : [[0, 'asc'], [2, 'asc'], [6, 'asc']],
-        dom: `<'row'<'col-12'B>> + <'row'<'col-sm-12'tr>> + <'row'<'col-md-12 col-lg-5'i><'col-md-12 col-lg-7'p>>`, // https://datatables.net/reference/option/dom#Styling
+        dom: "<'row'<'col-12'B>> + <'row'<'col-sm-12'tr>> + <'row'<'col-md-12 col-lg-5'i><'col-md-12 col-lg-7'p>>", // https://datatables.net/reference/option/dom#Styling
+        language: datatableTranslations,
         search: {
             regex: true, // Enable escaping of regular expression characters in the search term.
         },